move extpreimage invoice to channeldb
This moves the extpreimage invoice handling to the channeldb package,
where it really belongs, but in a separate file to still make
ongoing maintenance easier.
